requirements.txt
aiogram
aiohttp
python-dotenv
motor
yandex-ai-studio-sdk